Remove nested loop
Don't actually need a loop for each copy of the current card. We can
just add the number of copies to each subsequent card's count.

Also experimented between using the `entry` method and the `Entry` type
methods vs. the standard map `get` and `insert` methods. Line 9 of
`src/part2.rs` uses `entry` and `or_insert` to ensure that the first
card still has its count of 1 in the map. Using `get` and `unwrap_or(1)`
does not insert into the map, and we never account for the 1 copy of the
first card.

Lines 14-15 could also be replaced with:
`*card_counts.entry(card_number).or_insert(1) += num_copies;`. This also
gets the correct answer, while calling another `or_insert`. I'll leave
the lines as is as an example of using `get` and `insert`.