Create one shared `main.py` to run/test each day
Command runner accepts a command line arg for the day to be run/tested
and calls `main.py` with this value, which dynamically loads the modules
for that day.

Reading in the input files was being handled by each day's `main.py`
file, and is now in a shared `utils.py` file. This is also being used by
each days `test.py`, so that all file reading is controlled through
one function.